refactor Fragments
