+++ app/src/main/java/com/example/jetpacktrainning/ui/fragments/BaseFragment.kt
@@ -1,0 +1,45 @@
+package com.example.jetpacktrainning.ui.fragments
+
+import android.content.Context
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.hilt.navigation.fragment.hiltNavGraphViewModels
+import androidx.viewbinding.ViewBinding
+import com.example.jetpacktrainning.R
+import com.example.jetpacktrainning.ui.viewmodel.MainViewModel
+
+typealias Inflate<T> = (LayoutInflater, ViewGroup?, Boolean) -> T
+
+abstract class BaseFragment<VB: ViewBinding>(
+    private val inflate: Inflate<VB>
+) : Fragment() {
+
+    private var _binding: VB? = null
+    protected val binding get() = _binding!!
+
+    protected lateinit var _context: Context
+
+    protected val mainViewModel: MainViewModel by hiltNavGraphViewModels(R.id.nav_graph)
+
+    override fun onAttach(context: Context) {
+        super.onAttach(context)
+        _context = context
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = inflate.invoke(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
