+++ app/src/main/java/com/example/jetpacktrainning/ui/fragments/DetailFragment.kt
@@ -1,42 +1,17 @@
 package com.example.jetpacktrainning.ui.fragments
 
-import android.content.Context
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.hilt.navigation.fragment.hiltNavGraphViewModels
-import com.example.jetpacktrainning.R
+import androidx.compose.runtime.ExperimentalComposeApi
 import com.example.jetpacktrainning.databinding.CountryDetailLayoutBinding
 import com.example.jetpacktrainning.tools.Status
 import com.example.jetpacktrainning.tools.displayError
 import com.example.jetpacktrainning.tools.displayLoading
-import com.example.jetpacktrainning.ui.viewmodel.MainViewModel
+import dagger.hilt.android.AndroidEntryPoint
 
-class DetailFragment : Fragment() {
-
-    private var _myFragmentBinding: CountryDetailLayoutBinding? = null
-    private val binding
-        get() = _myFragmentBinding!!
-
-    private val mainViewModel: MainViewModel by hiltNavGraphViewModels(R.id.nav_graph)
-
-    private lateinit var _context: Context
-
-    override fun onAttach(context: Context) {
-        super.onAttach(context)
-        _context = context
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _myFragmentBinding = CountryDetailLayoutBinding.inflate(inflater, container, false)
-        return binding.root
-    }
+@ExperimentalComposeApi
+@AndroidEntryPoint
+class DetailFragment : BaseFragment<CountryDetailLayoutBinding>(CountryDetailLayoutBinding::inflate) {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -54,15 +29,6 @@ class DetailFragment : Fragment() {
                     displayLoading(_context)
                 }
             }
-
         }
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        //Fragments outlive their views.
-        // Make sure you clean up any references to the binding class instance
-        // in the fragment's onDestroyView() method.
-        _myFragmentBinding = null
     }
 }
